refactor: extract multipart form data matcher test helper function
