+++ grid/conversion/doc-to-text/maps/cloudmersive.test.ts
@@ -1,42 +1,8 @@
-import { RequestBodyMatcher } from 'nock/types';
-
-import { replaceBoundaryInMultipartFormDataBody } from '../../../../test/helpers/replace_boundary';
+import { multipartFormDataBoundaryMatcherHook } from '../../../../test/helpers/multipart_form_data_boundary_matcher_hook';
 import { docToTextTest } from './doc-to-text';
 
 docToTextTest('cloudmersive', {
-  beforeRecordingLoad: (
-    recordings: {
-      path: string;
-      method?: string | undefined;
-      body?: RequestBodyMatcher | undefined;
-    }[]
-  ) => {
-    recordings.forEach(
-      (recording: {
-        path: string;
-        method?: string | undefined;
-        body?: RequestBodyMatcher | undefined;
-      }) => {
-        if (
-          recording.method === 'POST' &&
-          recording.path === '/convert/autodetect/to/txt'
-        ) {
-          const boundaryReplacement = '---boundary replaced';
-          //Boundary contains random value, we are replacing it with constant to match the recording
-          const recordedBody = replaceBoundaryInMultipartFormDataBody(
-            boundaryReplacement,
-            recording.body
-          );
-          recording.body = (body: RequestBodyMatcher | undefined) => {
-            const requestBody = replaceBoundaryInMultipartFormDataBody(
-              boundaryReplacement,
-              body
-            );
-
-            return recordedBody === requestBody;
-          };
-        }
-      }
-    );
-  },
+  beforeRecordingLoad: multipartFormDataBoundaryMatcherHook((recording) => {
+    return recording.method === 'POST' && recording.path === '/convert/autodetect/to/txt';
+  })
 });
