+++ grid/recruitment/create-candidate/maps/breezy-hr.test.ts
@@ -1,4 +1,4 @@
-import { replaceBoundaryInMultipartFormDataBody } from '../../../../test/helpers/replace_boundary';
+import { multipartFormDataBoundaryMatcherHook } from '../../../../test/helpers/multipart_form_data_boundary_matcher_hook';
 import { createCandidateTest } from './create-candidate';
 import { createCandidateFeaturesTest } from './create-candidate-features';
 
@@ -9,31 +9,11 @@ createCandidateTest(
     invalid: '12345ab789d1dd',
   },
   {
-    beforeRecordingLoad: recordings => {
-      recordings.forEach(recording => {
-        if (
-          recording.method === 'POST' &&
-          /^\/v3\/company\/.+\/position\/.+\/candidate\/.+\/resume$/m.test(
-            recording.path
-          )
-        ) {
-          const boundaryReplacement = '---boundary replaced';
-          //resume upload contains multipart/form-data body with boundary with randomized value, we replace the boundary with constant in recorded body and matched body
-          const recordedBody = replaceBoundaryInMultipartFormDataBody(
-            boundaryReplacement,
-            recording.body
-          );
-          recording.body = (body: nock.RequestBodyMatcher) => {
-            const requestBody = replaceBoundaryInMultipartFormDataBody(
-              boundaryReplacement,
-              body
-            );
-
-            return recordedBody === requestBody;
-          };
-        }
-      });
-    },
+    beforeRecordingLoad: multipartFormDataBoundaryMatcherHook((recording) => {
+      return recording.method === 'POST' && /^\/v3\/company\/.+\/position\/.+\/candidate\/.+\/resume$/m.test(
+        recording.path
+      );
+    })
   }
 );
